# add benchmark and improve performance