@hnlq715 2020-02-18
I think we should just keep simple logic, and no duplicated path should be guaranteed by the caller.