@rvolosatovs 2020-02-18
I agree with this change given https://github.com/TheThingsIndustries/protoc-gen-fieldmask/pull/36#discussion_r380774100 is fixed.