@rvolosatovs 2020-02-24
Merged in https://github.com/TheThingsIndustries/protoc-gen-fieldmask/commit/321adaa1af73437df4fd792e1c58358714e1ddfb